Authentication of users completed in UserController

+++ app/Http/Controllers/User/UserController.php
@@ -9,7 +9,9 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
+use Symfony\Component\HttpFoundation\Response as ResponseAlias;
 
 class UserController extends Controller
 {
@@ -20,10 +22,33 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::all();
-        return response()->json($users);
+        $current_user = Auth::user();
+        if($current_user['is_admin']){
+            $users = User::all();
+            return response()->json($users, ResponseAlias::HTTP_OK);
+        } else {
+            return response()->json(['message' => 'You are not authorized to access this resource'], ResponseAlias::HTTP_UNAUTHORIZED);
+        }
     }
 
+    /**
+     * login.
+     *
+     * @param Request $request
+     * @return JsonResponse
+     */
+    public function login(Request $request)
+    {
+        $credentials = $request->only('email', 'password');
+
+        if (Auth::attempt($credentials)) {
+            $user = Auth::user();
+            $token = $user->createToken('MyApp')->plainTextToken;
+            return response()->json(['token' => $token], ResponseAlias::HTTP_OK);
+        } else {
+            return response()->json(['error' => 'Unauthorised'], ResponseAlias::HTTP_UNAUTHORIZED);
+        }
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -34,7 +59,7 @@ class UserController extends Controller
         try {
             $mail = $request->input('email');
 
-            if(!Str::contains($mail, '@erfan.ir')) {
+            if(!$mail == 'admin@admin.ir') {
                 $is_admin = false;
             } else {
                 $is_admin = true;
@@ -53,13 +78,18 @@ class UserController extends Controller
                 'updated_at' => Carbon::now(),
             ]);
 
-            return response()->json($user);
+            return response()->json(['user' => $user, 'token' => $user->createToken('MyApp')->plainTextToken], ResponseAlias::HTTP_OK);
 
         } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+            return response()->json(['error' => $e->getMessage()], ResponseAlias::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
+    public function profile()
+    {
+        $user = Auth::user();
+        return response()->json(['user' => $user], ResponseAlias::HTTP_OK);
+    }
     /**
      * Display the specified resource.
      *
@@ -67,37 +97,32 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        $user = User::query()->findOrFail($id);
-
-        return response()->json($user);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param User $user
-     * @return Response
-     */
-    public function edit(User $user)
-    {
-        //
+        $current_user = Auth::user();
+        if($current_user['is_admin']) {
+            $user = User::find($id);
+            return response()->json($user, ResponseAlias::HTTP_OK);
+        } else {
+            return response()->json(['error' => 'Unauthorised'], ResponseAlias::HTTP_UNAUTHORIZED);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param User $user
-     * @return Response
      */
-    public function destroy(Request $request)
+    public function destroy($id)
     {
         try {
-            $user = User::query()->findOrFail($request->input('id'));
-            $user->delete();
-
-            return response()->json(['success' => true]);
+            $current_user = Auth::user();
+            if($current_user['is_admin']) {
+                $user = User::find($id);
+                $user->delete();
+                return response()->json(['message' => 'User deleted successfully'], ResponseAlias::HTTP_OK);
+            } else {
+                return response()->json(['error' => 'Unauthorised'], ResponseAlias::HTTP_UNAUTHORIZED);
+            }
         } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+            return response()->json(['error' => $e->getMessage()], ResponseAlias::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 }
